CODA-56 - Implement handling parallel requests
