+++ handlers.h
@@ -1,3 +1,4 @@
+#include "utils.h"
 #include <arpa/inet.h>
 #include <stdarg.h>
 #include <stdio.h>
