+++ utils.h
@@ -4,40 +4,16 @@
 #include <stdlib.h>
 #include <string.h>
 
-#define MAX_CLIENTS 30
-#define CHUNK_SIZE 256
-
-void startTCPSrv(struct sockaddr_in *srvAddr, int *srvFd, unsigned int port) {
-  if ((*srvFd = socket(AF_INET, SOCK_STREAM, 0)) == 0) {
-    perror("server socket");
-    exit(EXIT_FAILURE);
-  }
-
-  srvAddr->sin_family = AF_INET;
-  srvAddr->sin_addr.s_addr = INADDR_ANY;
-  srvAddr->sin_port = htons(port);
-  memset(srvAddr->sin_zero, '\0', sizeof srvAddr->sin_zero);
-
-  if (bind(*srvFd, (struct sockaddr *)srvAddr, sizeof(*srvAddr)) < 0) {
-    perror("server binding");
-    exit(EXIT_FAILURE);
-  }
-
-  if (listen(*srvFd, MAX_CLIENTS) < 0) {
-    perror("server listening");
-    exit(EXIT_FAILURE);
-  }
-}
+#define HEADER_SEPARATOR " :\n\r"
 
 char *getHeaderVal(char *request, char *header) {
-  static const char headerSeps[] = " :\n\r";
   char *body;
   char *params;
 
   char *requestCpy = calloc(strlen(request) + 1, sizeof(char));
   strcpy(requestCpy, request);
 
-  body = strtok(requestCpy, headerSeps);
+  body = strtok(requestCpy, HEADER_SEPARATOR);
 
   while (1) {
     if (body == NULL) {
@@ -45,9 +21,9 @@ char *getHeaderVal(char *request, char *header) {
     }
 
     if (strcmp(header, body) == 0) {
-      return strtok(NULL, headerSeps);
+      return strtok(NULL, HEADER_SEPARATOR);
     }
-    body = strtok(NULL, headerSeps);
+    body = strtok(NULL, HEADER_SEPARATOR);
   }
 
   free(requestCpy);
